bot/requirements.txt
alpaca-py>=0.35.0
pandas>=2.2.0
ta>=0.11.0
supabase>=2.11.0
anthropic>=0.42.0
google-genai>=1.0.0
python-dotenv>=1.0.1
websockets>=14.0
httpx>=0.28.0
numpy>=2.0.0
